Introduced global dispatcher for add+remove events
The global event dispatcher is supposed to be more robust and flexible than
the previous event handling mechanism. Any event of a certain type runs
through the dispatcher that keeps a list of QueryCollections that are
interested in the events. Every collection can then decide for itself if it
should forward individual events or not.